Check for existence of cookbooks download path before wiping out its contents.

+++ agents/lib/instance/cook/executable_sequence.rb
@@ -255,11 +255,13 @@ module RightScale
 
       @audit.create_new_section('Retrieving cookbooks') unless @cookbooks.empty?
       audit_time do
-        # first, wipe out preexisting cookbooks.
-        Dir.foreach(@download_path) do |entry|
-          FileUtils.remove_entry_secure(File.join(@download_path, entry)) if entry =~ /\A\d+\Z/
+        # first, wipe out any preexisting cookbooks in the download path
+        if File.directory?(@download_path)
+          Dir.foreach(@download_path) do |entry|
+            FileUtils.remove_entry_secure(File.join(@download_path, entry)) if entry =~ /\A\d+\Z/
+          end
         end
-
+        
         counter = 0
 
         @cookbooks.each_with_index do |cookbook_sequence, i|
